requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
english-words==2.0.1
filelock==3.12.2
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
six==1.16.0
sympy==1.12
tifffile==2023.7.10
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
